requirements.txt
crewai
unstructured
pyowm
tools
python-dotenv
langchain-groq